Object Tracker added and Improvements
- Added Object tracker (video/camera)
- Merged Picture and Video recorder
- Added Interface to choose between Picture mode or Video mode
- More improvements to folder structure
- Changed Media files with royalty free pictures and video
- Minor changes to UI
