+++ Mod4.py
@@ -1,66 +1,26 @@
 import cv2
 from datetime import datetime
-def Mod4():
+
+def cctv():
     cap = cv2.VideoCapture(0)
 
-
-    right, left = "", ""
+    fourcc = cv2.VideoWriter_fourcc(*'XVID')
+    out = cv2.VideoWriter(f'CCTV/{datetime.now().strftime("%y.%m.%d - %H.%M")}.avi', fourcc,20.0,(640,480))
 
     while True:
-        _, frame1 = cap.read()
-        frame1 = cv2.flip(frame1, 1)
-        _, frame2 = cap.read()
-        frame2 = cv2.flip(frame2, 1)
+        _, frame = cap.read()
 
-        diff = cv2.absdiff(frame2, frame1)
+        cv2.putText(frame, 'RECORDING...', (50,50), cv2.FONT_HERSHEY_COMPLEX,
+                        0.6, (0,0,255), 2)
+        cv2.putText(frame, f'{datetime.now().strftime("%y.%m.%d - %H:%M:%S")}', (50,80), cv2.FONT_HERSHEY_COMPLEX,
+                        0.6, (255,255,255), 2)
+
+        out.write(frame)
         
-        diff = cv2.blur(diff, (5,5))
-        
-        gray = cv2.cvtColor(diff, cv2.COLOR_BGR2GRAY)
-        
-        _, threshd = cv2.threshold(gray, 40, 255, cv2.THRESH_BINARY)
-        
-        contr, _ = cv2.findContours(threshd, cv2.RETR_TREE, cv2.CHAIN_APPROX_SIMPLE)
-        
-        x = 300
-        if len(contr) > 0:
-            max_cnt = max(contr, key=cv2.contourArea)
-            x,y,w,h = cv2.boundingRect(max_cnt)
-            cv2.rectangle(frame1, (x, y), (x+w, y+h), (0,255,0), 2)
-            cv2.putText(frame1, "MOTION", (10,80), cv2.FONT_HERSHEY_SIMPLEX, 2, (0,255,0), 2)
-            
-        else:
-            cv2.putText(frame1, "NO-MOTION", (10,80), cv2.FONT_HERSHEY_SIMPLEX, 2, (0,0,255), 2)
-            
-        
-        if right == "" and left == "":
-            if x > 500:
-                right = True
-            
-            elif x < 200:
-                left = True
-                
-        elif right:
-                if x < 200:
-                    print("to left")
-                    x = 300
-                    right, left = "", ""
-                    cv2.imwrite(f"Visitors/In/{datetime.now().strftime('%y-%m-%d-%H-%M-%S')}.jpg", frame1)
-            
-        elif left:
-                if x > 500:
-                    print("to right")
-                    x = 300
-                    right, left = "", ""
-                    cv2.imwrite(f"Visitors/Out/{datetime.now().strftime('%y-%m-%d-%H-%M-%S')}.jpg", frame1)
-            
-            
-        
-        cv2.imshow("", frame1)
-        
-        k = cv2.waitKey(1)
-        
-        if k == 27:
+
+        cv2.imshow("CCTV", frame)
+
+        if cv2.waitKey(1) == 27:
             cap.release()
             cv2.destroyAllWindows()
-            break
+            break
